+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix Filters in User Admin
 * Fix Filters in Alias Admin
 * Improve Performance for Alias and Voucher Admin
+* Use Autocomplete for loading Users in Alias and Voucher Admin
 
 # 3.3.1 (2023.11.12)